pyproject.toml
[project]
name = "modal-dictation-exploration"
version = "0.1.0"
description = "Voice control software for Linux and macOS with modal dictation support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pasimple>=0.0.3",
    "pillow>=12.0.0",
    "pulsectl>=24.12.0",
    "pydantic>=2.12.5",
    "pygobject>=3.54.5",
    "pystray>=0.19.5",
]

[project.scripts]
modal-dictation-exploration = "modal_dictation_exploration.main:main"

[dependency-groups]
dev = [
    "polyfactory>=3.2.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
    "ty>=0.0.7",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.environment]
python-version = "3.12"
extra-paths = ["typings"]
python-platform = "linux"

[tool.ty.rules]
possibly-unresolved-reference = "error"
possibly-missing-attribute = "error"
possibly-missing-import = "error"
unused-ignore-comment = "warn"

[tool.ty.terminal]
error-on-warning = true

[tool.uv]
package = true
